refactor: global 패키지명을 system으로 변경
