+++ src/main/java/roomescape/member/controller/MemberController.java
@@ -4,8 +4,8 @@ import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.global.auth.annotation.Admin;
-import roomescape.global.dto.response.ApiResponse;
+import roomescape.system.auth.annotation.Admin;
+import roomescape.system.dto.response.ApiResponse;
 import roomescape.member.dto.MembersResponse;
 import roomescape.member.service.MemberService;
