+++ src/main/java/roomescape/reservation/controller/ReservationController.java
@@ -14,9 +14,9 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.global.auth.annotation.Admin;
-import roomescape.global.auth.annotation.MemberId;
-import roomescape.global.dto.response.ApiResponse;
+import roomescape.system.auth.annotation.Admin;
+import roomescape.system.auth.annotation.MemberId;
+import roomescape.system.dto.response.ApiResponse;
 import roomescape.reservation.dto.request.FilteredReservationRequest;
 import roomescape.reservation.dto.request.ReservationRequest;
 import roomescape.reservation.dto.response.MemberReservationsResponse;
