+++ src/main/java/roomescape/reservation/controller/ReservationTimeController.java
@@ -12,8 +12,8 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.global.auth.annotation.Admin;
-import roomescape.global.dto.response.ApiResponse;
+import roomescape.system.auth.annotation.Admin;
+import roomescape.system.dto.response.ApiResponse;
 import roomescape.reservation.dto.request.ReservationTimeRequest;
 import roomescape.reservation.dto.response.ReservationTimeResponse;
 import roomescape.reservation.dto.response.ReservationTimesResponse;
