+++ src/main/java/roomescape/reservation/service/ReservationService.java
@@ -7,11 +7,11 @@ import java.util.List;
 import java.util.Optional;
 import org.springframework.data.jpa.domain.Specification;
 import org.springframework.stereotype.Service;
-import roomescape.global.exception.error.ErrorType;
-import roomescape.global.exception.model.DataDuplicateException;
-import roomescape.global.exception.model.ForbiddenException;
-import roomescape.global.exception.model.NotFoundException;
-import roomescape.global.exception.model.ValidateException;
+import roomescape.system.exception.error.ErrorType;
+import roomescape.system.exception.model.DataDuplicateException;
+import roomescape.system.exception.model.ForbiddenException;
+import roomescape.system.exception.model.NotFoundException;
+import roomescape.system.exception.model.ValidateException;
 import roomescape.member.domain.Member;
 import roomescape.member.service.MemberService;
 import roomescape.reservation.domain.Reservation;
