+++ src/main/java/roomescape/system/auth/annotation/Admin.java
@@ -1,11 +1,11 @@
-package roomescape.global.auth.annotation;
+package roomescape.system.auth.annotation;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
-@Target(ElementType.PARAMETER)
+@Target(ElementType.METHOD)
 @Retention(RetentionPolicy.RUNTIME)
-public @interface MemberId {
+public @interface Admin {
 }
