+++ src/main/java/roomescape/system/auth/annotation/MemberId.java
@@ -1,11 +1,11 @@
-package roomescape.global.auth.annotation;
+package roomescape.system.auth.annotation;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
-@Target(ElementType.METHOD)
+@Target(ElementType.PARAMETER)
 @Retention(RetentionPolicy.RUNTIME)
-public @interface Admin {
+public @interface MemberId {
 }
