+++ src/main/java/roomescape/system/auth/interceptor/AdminInterceptor.java
@@ -1,4 +1,4 @@
-package roomescape.global.auth.interceptor;
+package roomescape.system.auth.interceptor;
 
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
@@ -7,11 +7,11 @@ import java.util.Arrays;
 import org.springframework.stereotype.Component;
 import org.springframework.web.method.HandlerMethod;
 import org.springframework.web.servlet.HandlerInterceptor;
-import roomescape.global.auth.annotation.Admin;
-import roomescape.global.auth.jwt.JwtHandler;
-import roomescape.global.exception.error.ErrorType;
-import roomescape.global.exception.model.ForbiddenException;
-import roomescape.global.exception.model.UnauthorizedException;
+import roomescape.system.auth.annotation.Admin;
+import roomescape.system.auth.jwt.JwtHandler;
+import roomescape.system.exception.error.ErrorType;
+import roomescape.system.exception.model.ForbiddenException;
+import roomescape.system.exception.model.UnauthorizedException;
 import roomescape.member.domain.Member;
 import roomescape.member.service.MemberService;
