+++ src/main/java/roomescape/system/auth/jwt/dto/TokenDto.java
@@ -1,4 +1,4 @@
-package roomescape.global.auth.jwt.dto;
+package roomescape.system.auth.jwt.dto;
 
 public record TokenDto(String accessToken, String refreshToken) {
 }
