+++ src/main/java/roomescape/system/auth/resolver/MemberIdResolver.java
@@ -1,4 +1,4 @@
-package roomescape.global.auth.resolver;
+package roomescape.system.auth.resolver;
 
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
@@ -9,10 +9,10 @@ import org.springframework.web.bind.support.WebDataBinderFactory;
 import org.springframework.web.context.request.NativeWebRequest;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.method.support.ModelAndViewContainer;
-import roomescape.global.auth.annotation.MemberId;
-import roomescape.global.auth.jwt.JwtHandler;
-import roomescape.global.exception.error.ErrorType;
-import roomescape.global.exception.model.UnauthorizedException;
+import roomescape.system.auth.annotation.MemberId;
+import roomescape.system.auth.jwt.JwtHandler;
+import roomescape.system.exception.error.ErrorType;
+import roomescape.system.exception.model.UnauthorizedException;
 
 @Component
 public class MemberIdResolver implements HandlerMethodArgumentResolver {
