+++ src/main/java/roomescape/system/config/WebMvcConfig.java
@@ -1,12 +1,12 @@
-package roomescape.global.config;
+package roomescape.system.config;
 
 import java.util.List;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import roomescape.global.auth.interceptor.AdminInterceptor;
-import roomescape.global.auth.resolver.MemberIdResolver;
+import roomescape.system.auth.interceptor.AdminInterceptor;
+import roomescape.system.auth.resolver.MemberIdResolver;
 
 @Configuration
 public class WebMvcConfig implements WebMvcConfigurer {
