+++ src/main/java/roomescape/system/dto/response/ApiResponse.java
@@ -1,4 +1,4 @@
-package roomescape.global.dto.response;
+package roomescape.system.dto.response;
 
 public record ApiResponse<T>(String message, T data) {
     private static final String SUCCESS_MESSAGE = "요청이 성공적으로 수행되었습니다.";
