+++ src/main/java/roomescape/theme/controller/ThemeController.java
@@ -13,8 +13,8 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.global.auth.annotation.Admin;
-import roomescape.global.dto.response.ApiResponse;
+import roomescape.system.auth.annotation.Admin;
+import roomescape.system.dto.response.ApiResponse;
 import roomescape.theme.dto.ThemeRequest;
 import roomescape.theme.dto.ThemeResponse;
 import roomescape.theme.dto.ThemesResponse;
