+++ src/test/java/roomescape/auth/service/AuthServiceTest.java
@@ -4,13 +4,12 @@ import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.context.annotation.Import;
 import roomescape.auth.dto.LoginRequest;
-import roomescape.global.auth.jwt.JwtHandler;
-import roomescape.global.auth.jwt.dto.TokenDto;
-import roomescape.global.exception.model.NotFoundException;
+import roomescape.system.auth.jwt.JwtHandler;
+import roomescape.system.auth.jwt.dto.TokenDto;
+import roomescape.system.exception.model.NotFoundException;
 import roomescape.member.domain.Member;
 import roomescape.member.domain.Role;
 import roomescape.member.domain.repository.MemberRepository;
