+++ src/test/java/roomescape/reservation/domain/ReservationTest.java
@@ -5,7 +5,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
-import roomescape.global.exception.model.ValidateException;
+import roomescape.system.exception.model.ValidateException;
 import roomescape.member.domain.Member;
 import roomescape.member.domain.Role;
 import roomescape.theme.domain.Theme;
