+++ src/test/java/roomescape/reservation/domain/ReservationTimeTest.java
@@ -3,7 +3,7 @@ package roomescape.reservation.domain;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import roomescape.global.exception.model.ValidateException;
+import roomescape.system.exception.model.ValidateException;
 
 class ReservationTimeTest {
