+++ src/test/java/roomescape/reservation/service/ReservationServiceTest.java
@@ -7,9 +7,9 @@ import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.context.annotation.Import;
 import org.springframework.test.context.jdbc.Sql;
 import org.springframework.test.context.jdbc.Sql.ExecutionPhase;
-import roomescape.global.exception.model.DataDuplicateException;
-import roomescape.global.exception.model.NotFoundException;
-import roomescape.global.exception.model.ValidateException;
+import roomescape.system.exception.model.DataDuplicateException;
+import roomescape.system.exception.model.NotFoundException;
+import roomescape.system.exception.model.ValidateException;
 import roomescape.member.domain.Member;
 import roomescape.member.domain.Role;
 import roomescape.member.domain.repository.MemberRepository;
