+++ src/test/java/roomescape/reservation/service/ReservationTimeServiceTest.java
@@ -3,13 +3,12 @@ package roomescape.reservation.service;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.context.annotation.Import;
 import org.springframework.test.context.jdbc.Sql;
 import org.springframework.test.context.jdbc.Sql.ExecutionPhase;
-import roomescape.global.exception.model.AssociatedDataExistsException;
-import roomescape.global.exception.model.DataDuplicateException;
+import roomescape.system.exception.model.AssociatedDataExistsException;
+import roomescape.system.exception.model.DataDuplicateException;
 import roomescape.member.domain.Member;
 import roomescape.member.domain.Role;
 import roomescape.member.domain.repository.MemberRepository;
